form-code
This .frm file is a form-code that gives you a .c-file you can include in your code in order to generate the Dirac-Part of the so-called "Mother-Matrix". This file uses the complete Dirac-structure of the pion with its 4 Dressingfunction (E-H).
The Local Kall is introduced to speed up the calculation of the pion mass in the Dyson_test_2_motherang.cpp by collecting powers of scalar products like (p.P, q.P, q.p etc.). Later the form optimization gives you an optimized output for further use in the Dyson_test_2_mohterang.cpp, for example.
format 04 is the highest level optimization possible.
This form code is designed to give you the .c-file for your code in a plug-and-play manner.

WARNING: There is still one more thing to do:
In the .c-output file, there are expressions like     w[1]=1/( - P_P*p_p + pow(P_p,2));
But my code cant use this expression in this way, because the numerator is an integer (although my format is float) and the denominator is a complex expression (because of the scalar product P.p). 
change all these fractions to     w[1]=1.0/(...);
and you're good to go!
CHEERS MATE